+++ src/features/manageProfileSettings/components/DefaultProfileOption/index.ts
@@ -1,0 +1,1 @@
+export { DefaultProfileOption } from './DefaultProfileOption';