+++ src/features/manageProfileSettings/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useNavigateToManageProfileSettings } from './useNavigateToManageProfileSettings';
+export { usePreferredProfiles } from './usePreferredProfiles';
 export { useProfileList } from './useProfileList';