+++ src/features/manageProfileSettings/index.ts
@@ -1,7 +1,9 @@
+export { CustomProfileToggle } from './components/CustomProfileToggle';
+export { DefaultProfileOption } from './components/DefaultProfileOption';
 export { ManageProfileSettingsControlPane } from './components/ManageProfileSettingsControlPane';
 export { ManageProfileSettingsControls } from './components/ManageProfileSettingsControls';
 export { ModalCloseProfileSettings } from './components/ModalCloseProfileSettings';
 export { ProfileSettings } from './components/ProfileSettings';
 export { ProfilesList } from './components/ProfilesList';
 
-export { useProfileList, useNavigateToManageProfileSettings } from './hooks';
+export { useProfileList, usePreferredProfiles, useNavigateToManageProfileSettings } from './hooks';